+++ CHANGELOG.md
@@ -8,6 +8,13 @@
 - Fix issue where `useAiChat()` didn't re-render correctly when chat title gets
   updated.
 
+### `@liveblocks/node`
+
+- Fix issue where `tenantId` was not being passed to the request when using
+  `Liveblocks.createRoom()`.
+
+- Add `comments:write` to the list of possible room permissions.
+
 ## v3.8.0
 
 ### `@liveblocks/client`